Revert "refactor(calendar): days-table"
This reverts commit 1c87babe06801d143b2cfe955b423cb2280fdd58.
