--- packages/calendar/src/components/day-button/index.ts
@@ -1,1 +1,0 @@
-export * from './Component';
